[ALS-8903] NHANES: age (#18)
* Fix concept node processing order
* Fix the order of metadata creation in DictionaryLoaderService
* Use `Interior` directly instead of null

The use of null is violating the `not null` constraint.

* Add new `INTERIOR` concept type and improve type handling
Introduced a new `INTERIOR` concept type and refactored `ConceptTypes` to better handle column metadata when determining concept types.

* Refactor concept type handling and add tests for intermediate nodes
Replaced string literals for concept types with `ConceptTypes` constants for consistency and readability. Added new test case to verify proper marking of non-leaf nodes as intermediate. Updated existing tests to align with the changes in concept type handling.